pyproject.toml
[tool.poetry]
name = "service-sls-process-de-para-account-ecd"
version = "0.1.0"
description = ""
authors = ["Elder Vivot <eldervivot@gmail.com>"]
readme = "README.md"
packages = [{include = "service_sls_process_de_para_account_ecd"}]

[tool.poetry.dependencies]
python = "^3.9"
aiohttp = "^3.8.4"
smart-open = "^6.3.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
flake8 = "^6.0.0"
autopep8 = "^2.0.2"
boto3 = "^1.26.123"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
